Add "Decentralized Graffiti" under Tools (#2)
While this tool is not limited to the use of the rocketpool community, I think it's a perfect fit. It doesn't improve the staking experience so it's just a fun thing, but if we choose to coordinate the drawing of a big *ss rocketpool logo or something, it could be seen as a form of advertisement.