Extracted 'Community FAQ' content to separate file
Currently, the current main page is a bit long, while we're not even live yet. So to shorten it's content without losing valuable information, I propose we put the Community FAQ contents into this separate file. When this change is accepted, we can adjust the main page to link to this one for community FAQ related content.