Added Community Tool 'Rocket Pool Gas Estimates'
This is a new version of https://github.com/isidorosp/awesome-rocketpool/tree/VGR-GIT-Added-Community-Tool-RocketPoolGasEstimates, which was the pull request that was requested before the rework that changed the subcategories/sections. (From: Article, Video ... -> Grouped by Content) It was easier to just make the change in the up to date main branch. The comments that blocked the initial pull request seemed void because the subcategories/sections were reworked.