Split up Tools (Official & Community) & Added Glossary
I've had some input (from Ken) that he couldn't find the RPL Calculator provided by the team, even though it was listed as a community tool. That's why this is my proposal; we have 3 main sources of content (Official, Press & Community) & if one of the subsections/categories is present under multiple sources of content, e.g. Explainer Series, then we add some context to it and move items to their correct place. In this PR this translates itself into adding more context to the existing 'Tools' subsection/category, so after this PR we have 'Community Tools' and 'Official Tools'. Under 'Official Tools' users will be able to find the official RPL calculator sheet.

As a minor change, I've also added the official glossary of RocketPool.